+++ package.json
@@ -14,6 +14,7 @@
     "@emotion/styled": "^11.11.0",
     "axios": "^1.6.5",
     "jwt-decode": "^4.0.0",
+    "moment": "^2.30.1",
     "next": "14.0.4",
     "react": "^18",
     "react-dom": "^18",
