+++ src/app/dashboard/header.tsx
@@ -1,6 +1,8 @@
 "use client";
 
 import { useState } from "react";
+import { useRouter } from "next/navigation";
+
 import {
   Box,
   Flex,
@@ -18,7 +20,6 @@ import {
 } from "@chakra-ui/react";
 import Logo from "./logo";
 import { useAppContext } from "@/context/authContext";
-import { useRouter } from "next/navigation";
 
 const MenuItem = ({
   children,
@@ -67,8 +68,9 @@ const MenuIcon = () => (
 const Header = () => {
   const [show, setShow] = useState(false);
   const toggleMenu = () => setShow(!show);
+  const { user, setIsAuth, setUser } = useAppContext();
+
   const { push } = useRouter();
-  const { user, setIsAuth } = useAppContext();
   return (
     <Flex
       as="nav"
@@ -99,9 +101,9 @@ const Header = () => {
           direction={["column", "row", "row", "row"]}
           pt={[4, 4, 0, 0]}
         >
+          <MenuItem to="/landing">Landing</MenuItem>
           <MenuItem to="/home">Home</MenuItem>
-          <MenuItem to="/experience">Experience </MenuItem>
-          <MenuItem to="/portfolio">Portfolio</MenuItem>
+          <MenuItem to="/dashboard">Dashboard </MenuItem>
           <MenuItem to="/faetures">Features </MenuItem>
           <MenuItem to="/pricing">Pricing </MenuItem>
         </Flex>
@@ -140,7 +142,7 @@ const Header = () => {
                 <Center>
                   <Box>
                     <Text fontWeight="bold">
-                      {user.name}
+                      {user && user.name}
                       <br />
                       <Badge ml="1" colorScheme="green">
                         ACTIVE - Freelancer
@@ -152,7 +154,9 @@ const Header = () => {
                       colorScheme="red"
                       onClick={() => {
                         setIsAuth(false);
-                        push("/landing");
+                        localStorage.removeItem("user");
+                        setUser({});
+                        push("/");
                       }}
                     >
                       Sign out
