--- src/app/portfolio/index.css
@@ -1,241 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Poppins:wght@300&display=swap");
-.blob {
-  animation: move 4s infinite;
-  animation-delay: 2s;
-}
-* {
-  font-family: "Poppins", sans-serif;
-  scroll-behavior: smooth;
-  text-align: justify;
-}
-
-::-webkit-scrollbar {
-  display: none;
-}
-::-webkit-scrollbar-track {
-  background: #f8eded;
-}
-::-webkit-scrollbar-thumb {
-  background: #ff0080;
-}
-::-webkit-scrollbar-thumb:hover {
-  background: #ff4ca5;
-}
-
-.reveal {
-  position: relative;
-  transform: translateY(150px);
-  opacity: 0;
-  transition: 1s all ease;
-}
-
-.reveal.active {
-  transform: translateY(0);
-  opacity: 1;
-}
-@keyframes move {
-  0% {
-    transform: scale(0.95);
-    box-shadow: 0 0 0 0 rgba(0, 0, 0, 0.7);
-  }
-
-  70% {
-    transform: scale(1);
-    box-shadow: 0 0 0 10px rgba(0, 0, 0, 0);
-  }
-
-  100% {
-    transform: scale(0.95);
-    box-shadow: 0 0 0 0 rgba(0, 0, 0, 0);
-  }
-}
-.main {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  font-family: sans-serif;
-  flex-wrap: wrap;
-}
-
-.card {
-  position: relative;
-  width: 400px;
-  height: 650px;
-  box-shadow: 0 0 13px 0px rgba(0, 0, 0, 0.3);
-  transition: 0.5s;
-  overflow: hidden;
-  margin: 1.15rem;
-}
-
-.card img {
-  width: 400px;
-  height: 250px;
-  transition: 2.5s;
-}
-.card .card-body {
-  padding: 0.5rem 0.85rem 1rem;
-  height: 0;
-}
-
-.card .card-body h6 {
-  margin: 0.5rem 0;
-  font-size: 1rem;
-  font-weight: 600;
-  color: crimson;
-  letter-spacing: 3px;
-  font-family: "Poppins", sans-serif;
-}
-
-.card .card-body p {
-  text-align: justify;
-  font-size: 1rem;
-  font-weight: 550;
-  line-height: 25px;
-  font-family: "Poppins", sans-serif;
-}
-
-.card .card-body div {
-  font-family: "Poppins", sans-serif;
-  text-align: right;
-  width: 100%;
-}
-
-.card .card-body div a {
-  position: relative;
-  text-decoration: none;
-  color: #526ffd;
-  padding: 0.5rem;
-  text-align: center;
-  z-index: 999;
-  font-size: 0.85rem;
-}
-
-.card .card-body div a::before {
-  position: absolute;
-  content: "";
-  width: 13px;
-  height: 13px;
-  background: #526ffd;
-  border-radius: 300px;
-  left: -13px;
-  top: 3px;
-  z-index: -1;
-  transition: 2.3s;
-}
-
-.card .card-body div a:hover {
-  color: #fff;
-}
-
-.card .card-body div a:hover::before {
-  width: 100%;
-  height: 88%;
-  left: 0;
-}
-
-@media screen and (max-width: 768px) {
-  .card {
-    width: 100%;
-    height: 670px;
-    margin-top: 25px;
-  }
-  .card img {
-    width: 100%;
-    height: 220px;
-  }
-  .card .card-body {
-    padding: 1.5rem 0.85rem 1rem;
-    height: 0;
-  }
-  .card .card-body h6 {
-    margin: 0.5rem 0;
-    font-size: 1rem;
-    font-weight: 600;
-    color: crimson;
-    letter-spacing: 3px;
-    font-family: "Poppins", sans-serif;
-  }
-  .card .card-body p {
-    text-align: justify;
-    font-size: 1rem;
-    line-height: 25px;
-    font-family: "Poppins", sans-serif;
-  }
-  .card .card-body div {
-    font-family: "Poppins", sans-serif;
-    text-align: right;
-    width: 100%;
-  }
-  .card .card-body div a {
-    position: relative;
-    text-decoration: none;
-    color: #526ffd;
-    padding: 0.5rem;
-    text-align: center;
-    z-index: 999;
-    font-size: 0.85rem;
-  }
-  .card .card-body div a::before {
-    position: absolute;
-    content: "";
-    width: 13px;
-    height: 13px;
-    background: #526ffd;
-    border-radius: 300px;
-    left: -13px;
-    top: 3px;
-    z-index: -1;
-    transition: 2.3s;
-  }
-  .card .card-body div a:hover {
-    color: #fff;
-  }
-  .card .card-body div a:hover::before {
-    width: 100%;
-    height: 88%;
-    left: 0;
-  }
-}
-
-.dot {
-  width: 0.7rem;
-  height: 0.7rem;
-  background-color: rgb(83, 197, 63);
-  border-radius: 50%;
-  display: inline-block;
-  animation: blink 1s infinite;
-  margin: 0 0.1rem 0 0.1rem;
-}
-.blink {
-  animation: blink 1s infinite;
-}
-/* --- Animation --- */
-
-/* Define animation keyframes */
-@keyframes blink {
-  0%,
-  100% {
-    opacity: 0;
-  }
-  50% {
-    opacity: 1;
-  }
-}
-
-/* Minimal example */
-.dot--basic {
-  animation: blink 2s infinite;
-}
-
-/* Run animation once */
-.dot--once {
-  animation: blink 2s 1;
-  /* animation-iteration-count: 1; */
-}
-
-/* Wait 4s before running the animation */
-.dot--delayed {
-  animation: blink 2s infinite 4s;
-  /* animation-delay: 4s; */
-}
